+++ crates/stackable-operator/src/kvp/label/value.rs
@@ -1,7 +1,7 @@
 use std::{fmt::Display, ops::Deref, str::FromStr, sync::LazyLock};
 
 use regex::Regex;
-use snafu::{ensure, Snafu};
+use snafu::{Snafu, ensure};
 
 use crate::kvp::Value;
 
@@ -56,12 +56,9 @@ impl FromStr for LabelValue {
     fn from_str(input: &str) -> Result<Self, Self::Err> {
         // The length of the value cannot exceed 63 characters, but can be
         // empty
-        ensure!(
-            input.len() <= LABEL_VALUE_MAX_LEN,
-            ValueTooLongSnafu {
-                length: input.len()
-            }
-        );
+        ensure!(input.len() <= LABEL_VALUE_MAX_LEN, ValueTooLongSnafu {
+            length: input.len()
+        });
 
         // The value cannot contain non-ascii characters
         ensure!(input.is_ascii(), ValueNotAsciiSnafu);