+++ crates/stackable-telemetry/src/tracing/settings/file_log.rs
@@ -2,6 +2,9 @@
 
 use std::path::PathBuf;
 
+/// Re-export to save the end crate from an extra import.
+pub use tracing_appender::rolling::Rotation;
+
 use super::{Settings, SettingsToggle};
 
 /// Configure specific settings for the File Log subscriber.
@@ -18,6 +21,15 @@ pub enum FileLogSettings {
 
         /// Path to directory for log files.
         file_log_dir: PathBuf,
+
+        /// Log rotation frequency.
+        rotation_period: Rotation,
+
+        /// Suffix for log filenames.
+        filename_suffix: String,
+
+        /// Keep the last `n` files on disk.
+        max_log_files: Option<usize>,
     },
 }
 
@@ -38,14 +50,32 @@ impl SettingsToggle for FileLogSettings {
 pub struct FileLogSettingsBuilder {
     pub(crate) common_settings: Settings,
     pub(crate) file_log_dir: PathBuf,
+    pub(crate) rotation_period: Rotation,
+    pub(crate) filename_suffix: String,
+    pub(crate) max_log_files: Option<usize>,
 }
 
 impl FileLogSettingsBuilder {
+    /// Set file rotation period.
+    pub fn with_rotation_period(mut self, rotation_period: impl Into<Rotation>) -> Self {
+        self.rotation_period = rotation_period.into();
+        self
+    }
+
+    /// Set maximum number of log files to keep.
+    pub fn with_max_files(mut self, max_log_files: impl Into<Option<usize>>) -> Self {
+        self.max_log_files = max_log_files.into();
+        self
+    }
+
     /// Consumes self and returns a valid [`FileLogSettings`] instance.
     pub fn build(self) -> FileLogSettings {
         FileLogSettings::Enabled {
             common_settings: self.common_settings,
             file_log_dir: self.file_log_dir,
+            rotation_period: self.rotation_period,
+            filename_suffix: self.filename_suffix,
+            max_log_files: self.max_log_files,
         }
     }
 }
@@ -76,11 +106,16 @@ mod test {
                 default_level: LevelFilter::DEBUG,
             },
             file_log_dir: PathBuf::from("/logs"),
+            rotation_period: Rotation::HOURLY,
+            filename_suffix: "tracing-rs.log".to_owned(),
+            max_log_files: Some(6),
         };
         let result = Settings::builder()
             .with_environment_variable("hello")
             .with_default_level(LevelFilter::DEBUG)
-            .file_log_settings_builder(PathBuf::from("/logs"))
+            .file_log_settings_builder(PathBuf::from("/logs"), "tracing-rs.log")
+            .with_rotation_period(Rotation::HOURLY)
+            .with_max_files(6)
             .build();
 
         assert_eq!(expected, result);