+++ crates/stackable-versioned/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.7.1] - 2025-04-02
+
+### Fixed
+
+- Correctly emit generic type parameter defaults in enum/struct definition blocks ([#991]).
+
+[#991]: https://github.com/stackabletech/operator-rs/pull/991
+
 ## [0.7.0] - 2025-03-17
 
 ### Changed